requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langgraph==0.0.26
langchain==0.1.20
langchain-nvidia-ai-endpoints==0.1.6
pydantic==2.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
